pyproject.toml
[tool.poetry]
name = "personal-assistant-web-application-django"
version = "0.1.0"
description = ""
authors = ["kirienkovitaliy <kirienkovitaliy1989@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.6"
django-cloudinary-storage = "^0.3.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
sphinx = ">=1.6, <7"
sphinx-rtd-theme = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
